+++ src/main/java/com/auca/kigalimotaricoperative/Controller/MotariController.java
@@ -1,7 +1,9 @@
-package com.auca.kigalimotaricoperative.controller;
+package com.auca.kigalimotaricoperative.Controller;
 
+import com.auca.kigalimotaricoperative.model.Admin;
 import com.auca.kigalimotaricoperative.model.Motari;
 import com.auca.kigalimotaricoperative.service.MotariService;
+import jakarta.servlet.http.HttpSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -16,13 +18,20 @@ public class MotariController {
     private MotariService motariService;
 
     @GetMapping("/motari")
-    public String allMotaris(Model model) {
-        Motari motari = new Motari();
-        List<Motari> motaris = motariService.getAllMotaris();
-        model.addAttribute("motari", motari);
-        model.addAttribute("motaris", motaris);
+    public String allMotaris(Model model, HttpSession session) {
+        Admin collectAdmin=(Admin)  session.getAttribute("adminLogedIn");
+        if(collectAdmin != null){
+            model.addAttribute("collectAdmin",collectAdmin);
+            Motari motari = new Motari();
+            List<Motari> motaris = motariService.getAllMotaris();
+            model.addAttribute("motari", motari);
+            model.addAttribute("motaris", motaris);
 
-        return "motari";
+            return "motari";
+        }else {
+            return "redirect:/";
+        }
+
     }
 
     @PostMapping("/newMotari")
@@ -32,22 +41,45 @@ public class MotariController {
     }
 
     @GetMapping("/motariForm")
-    public String showCreateForm(Model model) {
-        model.addAttribute("motari", new Motari());
-        return "motariForm";
+    public String showCreateForm(Model model,HttpSession session) {
+        Admin collectAdmin=(Admin)  session.getAttribute("adminLogedIn");
+        if(collectAdmin != null){
+            model.addAttribute("collectAdmin",collectAdmin);
+            model.addAttribute("motari", new Motari());
+            return "motariForm";
+        }else {
+            return "redirect:/";
+        }
+
     }
 
     @GetMapping("/motariEdit/{motariId}")
-    public String showEdit(@PathVariable String motariId, Model model) {
-        Motari motari = motariService.getMotariById(motariId);
+    public String showEdit(@PathVariable Integer motariId, Model model,HttpSession session) {
+        Admin collectAdmin=(Admin)  session.getAttribute("adminLogedIn");
+        if(collectAdmin != null){
+            model.addAttribute("collectAdmin",collectAdmin);
+            Motari motari = motariService.getMotariById(motariId);
 
-        model.addAttribute("motari", motari);
-        return "redirect:/motari";
+            model.addAttribute("motari", motari);
+            return "redirect:/motari";
+        }else {
+            return "redirect:/";
+        }
+
     }
 
     @GetMapping("/motariDelete/{motariId}")
-    public String deleteMotari(@PathVariable String motariId) {
-        motariService.deleteMotari(motariId);
-        return "redirect:/motari";
+    public String deleteMotari(@PathVariable Integer motariId,HttpSession session,Model model) {
+        Admin collectAdmin=(Admin)  session.getAttribute("adminLogedIn");
+        if(collectAdmin != null){
+            model.addAttribute("collectAdmin",collectAdmin);
+            Motari motari = motariService.getMotariById(motariId);
+
+            motariService.deleteMotari(motariId);
+            return "redirect:/motari";
+        }else {
+            return "redirect:/";
+        }
+
     }
 }
